Fix a race condition in RSASigner::Impl::sign()
The sign() method of RSASigner::Impl is const and thread-safe by itself
however it uses an instance of RandomPool which is not thread safe. As a
result concurrent calls to sign() lead to data races inside the
RandomPool.

This patch fixes this by making the RandomPool instance thread local.
This way each thread works with its own object and there are no races.

RandomPool is wrapped in a new type because an adequate default
constructor needs to be provided. Otherwise the object initialisation
becomes complicated.